+++ lib/libchen/CHANGELOG.md
@@ -11,9 +11,13 @@ All notable changes to this project will be documented in this file(see http://k
 - sys: retrieve current executable path.
 - sys: retrieve current process id.
 - sys: kill a specific process.
+- proc: check existence of process.
 - regex: add methods support pass std::regex object.
 - regex: add expand method that can use submatch to expand a string.
 - threadpool: user can retrieve the threads count.
+
+### Changed
+- sys: move process related methods to proc class.
 
 ### Fixed
 - ini: fix parse error if comment between two properties.